requirements.txt
\dash==2.8.1
dash_ag_grid==2.0.0
dash_bootstrap_components==1.0.0
dash_core_components==2.0.0
dash_extensions==0.0.71
dash_html_components==2.0.0
dash_labs==1.0.1
Flask==2.0.2
gunicorn==20.0.4
keras==2.7.0
loopgpt==0.0.15
matplotlib==3.4.3
mplfinance==0.12.7a17
numpy==1.23.2
pandas==1.4.3
pandas_market_calendars==3.4
plotly==5.13.0
pylightxl==1.58
pymongo==3.12.1
pyorbital==1.6.1
python_dateutil==2.8.2
Requests==2.31.0
scikit_learn==1.0.1
xmltodict==0.12.0
yfinance==0.1.70
